requirements.txt
albumentations==1.4.8
fastai==2.7.15
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.10.0.82
pandas==2.0.3
Pillow==10.3.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
spacy==3.7.5
requests==2.32.3
